io_out.cc fixes related to long MAXHANDLE

+++ gnucap/src/io_out.cc
@@ -56,7 +56,7 @@ const char* octal(int x)
  */
 OMSTREAM & OMSTREAM::tab(unsigned count)
 {
-  for (int ii=0, mm=1;   ii<=MAXHANDLE;   ++ii, mm<<=1) {
+  for (long ii=0, mm=1;   ii<=MAXHANDLE;   ++ii, mm<<=1) {
     if (_mask & mm) {
       OMSTREAM this_file(_mask & mm);
       if (_cpos[ii] > count) {
@@ -125,7 +125,7 @@ OMSTREAM & OMSTREAM::operator<<(const char *str)
   }
 
   /* auto line break, with a '+' to continue. */
-  for (int ii=0, mm=1;   ii<=MAXHANDLE;   ++ii, mm<<=1) {
+  for (long ii=0, mm=1;   ii<=MAXHANDLE;   ++ii, mm<<=1) {
     if ((_mask & mm)
 	&& (sl+_cpos[ii]) >= OPT::outwidth
 	&& _cpos[ii] != 0) {
@@ -188,7 +188,7 @@ OMSTREAM & OMSTREAM::operator<<(char chr)
   }else{
   }
   
-  for (int ii=0, mm=1;   ii<=MAXHANDLE;   ++ii, mm<<=1) {
+  for (long ii=0, mm=1;   ii<=MAXHANDLE;   ++ii, mm<<=1) {
     if (_mask & mm) {
       assert(IO::stream[ii]);
       if (chr=='\b') {untested();
@@ -229,7 +229,7 @@ bool OMSTREAM::writeout(char chr)
 		error(bDANGER, "internal error: out to stdin\n");
 	}else{
 	}
-	for (int ii=0, mm=1;   ii<=MAXHANDLE;   ++ii, mm<<=1) {
+	for (long ii=0, mm=1;   ii<=MAXHANDLE;   ++ii, mm<<=1) {
 		if (_mask & mm) {
 			fputc(chr,IO::stream[ii]);
 			//fflush(IO::stream[ii]);
